move utility function arround
