+++ index.js
@@ -212,7 +212,7 @@ Hermes.prototype.getQueues = function () {
 Hermes.prototype.publish = function (queueName, data) {
   /*jshint maxcomplexity:7 */
   debug('hermes publish', queueName, data);
-  if (!Hermes._doesQueueExist(this._opts.queues, queueName) && !this._eventJobs.isPublishEvent(queueName)) {
+  if (!assertOpts.doesQueueExist(this._opts.queues, queueName) && !this._eventJobs.isPublishEvent(queueName)) {
     throw new Error('attempting to publish to invalid queue: '+queueName);
   }
   if (typeof data === 'string' || data instanceof String || data instanceof Buffer) {
@@ -237,7 +237,7 @@ Hermes.prototype.publish = function (queueName, data) {
  */
 Hermes.prototype.subscribe = function (queueName, handler) {
   debug('hermes subscribe', queueName);
-  if (!Hermes._doesQueueExist(this._opts.queues, queueName) && !this._eventJobs.isSubscribeEvent(queueName)) {
+  if (!assertOpts.doesQueueExist(this._opts.queues, queueName) && !this._eventJobs.isSubscribeEvent(queueName)) {
     throw new Error('attempting to subscribe to invalid queue: ' + queueName);
   }
   if (handler.length < 2) {
@@ -259,7 +259,7 @@ Hermes.prototype.subscribe = function (queueName, handler) {
  */
 Hermes.prototype.unsubscribe = function (queueName, handler, cb) {
   debug('hermes unsubscribe', queueName);
-  if (!Hermes._doesQueueExist(this._opts.queues, queueName) && !this._eventJobs.isSubscribeEvent(queueName)) {
+  if (!assertOpts.doesQueueExist(this._opts.queues, queueName) && !this._eventJobs.isSubscribeEvent(queueName)) {
     throw new Error('attempting to unsubscribe from invalid queue: ' + queueName);
   }
   this.emit('unsubscribe', queueName, handler, cb);
@@ -378,19 +378,6 @@ Hermes._normalizeQueues = function (queues) {
   return queues.map(Hermes._normalizeQueue)
 }
 
-/**
- * Check if queue with provided `name` exists in the array of queueDefs
- * @param {Array} array of nromalized queueDefs
- * @param {String} queueName to check for existence
- * @param {Array} array of mixed queueNames or queueDefs
- * @return {Boolean} true if queue with `name` exists in the array of queueDefs
- */
-Hermes._doesQueueExist = function (queues, name) {
-  var result = queues.filter(function (queue) {
-    return queue.name === name
-  })
-  return result.length > 0
-}
 
 /**
  * Assert queue with provided name and potentially options
