+++ lib/assert-opts.js
@@ -22,6 +22,19 @@ var checkQueueDef = function (queueDef) {
 }
 
 /**
+ * Check if queue with provided `name` exists in the array of queueDefs
+ * @param {Array} array of nromalized queueDefs
+ * @param {String} queueName to check for existence
+ * @param {Array} array of mixed queueNames or queueDefs
+ * @return {Boolean} true if queue with `name` exists in the array of queueDefs
+ */
+var doesQueueExist = function (queues, name) {
+  var result = queues.filter(function (queue) {
+    return queue.name === name
+  })
+  return result.length > 0
+}
+/**
  * Assert valid Hermes configuration options
  * @throws
  * @param {Object} opts
@@ -69,3 +82,4 @@ module.exports = function (opts) {
 
 
 module.exports._checkQueueDef = checkQueueDef
+module.exports.doesQueueExist = doesQueueExist
