+++ lib/event-jobs.js
@@ -2,10 +2,11 @@
 
 module.exports = EventJobs;
 
-var async = require('async');
-var debug = require('debug')('hermes:lib:events');
+var async = require('async')
 
-var Hermes = require('../index')
+var assertOpts = require('./assert-opts')
+var debug = require('debug')('hermes:lib:events')
+
 
 /**
  * Used to handle all event operations
@@ -43,7 +44,7 @@ EventJobs.prototype.setChannel = function (ch) {
  */
 EventJobs.prototype.isPublishEvent = function (eventName) {
   debug('isPublishEvent', eventName);
-  return Hermes._doesQueueExist(this._publishedEvents, eventName);
+  return assertOpts.doesQueueExist(this._publishedEvents, eventName);
 };
 
 /**
@@ -94,7 +95,7 @@ EventJobs.prototype.publish = function (eventName, data, routingKey) {
  */
 EventJobs.prototype.isSubscribeEvent = function (eventName) {
   debug('isSubscribeEvent', eventName);
-  return Hermes._doesQueueExist(this._subscribedEvents, eventName);
+  return assertOpts.doesQueueExist(this._subscribedEvents, eventName);
 };
 
 /**
