+++ test/unit/lib/assert-opts.js
@@ -254,4 +254,26 @@ describe('assertOpts', function () {
       done()
     })
   })
+
+  describe('_doesQueueExist', function () {
+    it('should return true if queue exists', function (done) {
+      var queues = [
+        { name: 'a', opts: { durable: true } },
+        { name: 'b', opts: { durable: true } },
+        { name: 'c', opts: { durable: true } }
+      ]
+      expect(assertOpts.doesQueueExist(queues, 'b')).to.be.true()
+      done()
+    })
+
+    it('should return false if queue does not exist', function (done) {
+      var queues = [
+        { name: 'a', opts: { durable: true } },
+        { name: 'b', opts: { durable: true } },
+        { name: 'c', opts: { durable: true } }
+      ]
+      expect(assertOpts.doesQueueExist(queues, 'd')).to.be.false()
+      done()
+    })
+  })
 });
